fix: found a circular reference
