+++ tools/emissions_factors_sources/parametric_knowledge.py
@@ -1,4 +1,4 @@
-from tools.emissions_factors import EFState, EmissionsFactor
+from tools.emissions_factors_state import EFState, EmissionsFactor
 from langchain_openai import ChatOpenAI
 
 def parametric_knowledge_ef_finder(state: EFState):
